# docs/swagger.yaml
basePath: /
definitions:
...
  dto.ErrResponse:
    properties:
      error:
        type: string
      module:
        type: string
      type:
        type: string
    type: object
...
  dto.UserSearch:
    properties:
      avatar_url:
        type: string
      display_name:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
  dto.UserSearchResponse:
    properties:
      users:
        items:
          $ref: '#/definitions/dto.UserSearch'
        type: array
    type: object
info:
  contact: {}
  description: Realtime chat application
  title: Recho
  version: 0.0.0
paths:
...
  /api/user/search:
    get:
      consumes:
      - application/json
      description: Returns similar usernames with some info
      parameters:
      - description: username
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserSearchResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrResponse'
      security:
      - BearerAuth: []
      summary: Search in users by id
      tags:
      - user
securityDefinitions:
  CookieAuth:
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"